Project1:Arithmetic Arrange Continue..
basically what i did was finished the logic its working right added the validation of numbers with Error Message, Printed the Output with True and Optional Argument. What took time was not the logic but again how we convert the data structure of the output of a one line and thinking about it in a specific manner. Also did the total and prited the same with the output. The project is taking time, its not the logic Its about how we convert things in python like from a list to integer and how do we access a particual elements in a different data structure. The project still continues... Have to finish all the validation so that user gets a good feedback when things go wrong