Continue.. Project Arithmetic Arranger FCC
did all the logic and applied all the error printing when input is wrong. Now the only part left is the output formate Rules. Learned so much about how to use try and except. Also learned we can get the output by sys.exit But have to consider if your code gets existed with exit_code 0 -- otherwise if any other code.. it does not call successful termination of the code or programme.